requirements.txt
ib_insync==0.9.86
pandas==2.1.0
python-dotenv==1.0.0
schedule==1.2.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
yfinance==0.2.50
